chore: move jest-solid to testutils
It's not possible to fully test these, and they are only used as helpers in tests, not as part of application code
